requirements.txt
pandas
nltk
google-generativeai
python-dotenv 
google-genai